refactor: Step1 -> Step2 폴더 이동 #48
