+++ src/app/(mood)/join/components/FunnelContext.tsx
@@ -8,12 +8,12 @@ export type FunnelContextValue = Pick<
   'setStep' | 'currentStep' | 'Funnel' | 'prevStep' | 'nextStep'
 >;
 
-type Steps = '1' | '2' | '3' | '4' | '5';
+type Steps = '1' | '2' | '3' | '4' | '5' | '6';
 
 const FunnelContext = createContext<FunnelContextValue | null>(null);
 
 export default function FunnelProvider({ children }: PropsWithStrictChildren) {
-  const { setStep, currentStep, Funnel, prevStep, nextStep } = useFunnel<Steps>('1', '5');
+  const { setStep, currentStep, Funnel, prevStep, nextStep } = useFunnel<Steps>('1', '6');
 
   const memoizedValue = useMemo(
     () => ({
