+++ src/app/(mood)/join/funnels/step2/components/BodyTypeSection.tsx
